Add comments and move modpow to separate file
